(chore: core) updated fibonacci halo2 example to be in line with interface change
