+++ mopro-core/examples/halo2/fibonacci/src/circuit.rs
